Extended what-is command
Extended the what-is command so that it gives the short description of
every tile moved over by the cursor, as well as giving the full
functionality if a tile is selected. The far-look command still only
gives information on a tile when selected, not as the cursor moves.
Haven't decided yet if this is the best way to do things,
interface-wise.