Far look replaced by tile explore
Replace the functionality of the far look command (;) to let you explore
tiles, getting a short description of their contents as you move through
them. Could be extracted out into its own function, but mostly the same
as existing function, and the far look functionality all exists still in
the what is command, so it seems a decent tradeoff to make sure the tile
explore command is readily accessible.