Move interrupts for pickup
The pickup now checks if the item has been moved or picked up and stops.
Also deals with vague interruption when nomul has cleared multi, which
seems to happen whenever the most obvious path to the point is blocked,
such as by your pet. Annoyingly this does happen with the move command
as well, it's just sucky pathing.