Changed the travel to command ('_') to give menu of features
Changed the travel command ('_') so that it gives a menu of all the seen
dungeon features (stairs, ladders, fountains, thrones, sinks, graves,
and altars) with their relative location, and the player can select one
and travel there. If they say no to the prompt it acts like it did
before.