Changes to pickup - list location, use 'seen' instead of rooms
Updated the room pickup so that items have their relative location
listed. Also, instead of listing items in the room, list items that can
be seen. This has its own problems, since in a dark room you can't 'see'
items even if they are on the screen, but it fixes the hallway and
dwarven mine. Might be better to use glyphs?? Then it would account for
items that have been seen in the past, but has its own problems since
it's even more likely than usual that the item has been moved, in which
case I'd need to figure out when you 'realize' that and stop moving to
the square.