'v' to view things around you, replace #listmon, do when entering room.
Generalized out the #listmon function to instead list all visible
monsters, objects, and dungeon features. Then remapped this control to
be used with the 'v' key, which meant shifting the 'version' command to
be 'V', and the history command which used to be mapped to 'V' to be an
extended #history command. Updated some docs to support this, though
Guidebook still has to be updated. This command will also play any time
you enter a room. The extended command #listmon is now removed.